Minor editorial changes
Made some changes to the ordering of content so that it flows better. Changed a bit of the wording.  No change to the analysis itself was made, except that the decision tree model was using the full data set and not the training data. Interestingly, this made little difference to the results.